PoC for removing Rich from container requirements
