+++ modal/__init__.py
@@ -9,7 +9,6 @@ if sys.version_info[:2] >= (3, 13):
 from modal_version import __version__
 
 try:
-    from ._output import enable_output
     from ._tunnel import Tunnel, forward
     from .app import App, Stub
     from .client import Client
@@ -22,6 +21,7 @@ try:
     from .image import Image
     from .mount import Mount
     from .network_file_system import NetworkFileSystem
+    from .output import enable_output
     from .partial_function import asgi_app, batched, build, enter, exit, method, web_endpoint, web_server, wsgi_app
     from .proxy import Proxy
     from .queue import Queue
