+++ modal/_output.py
@@ -704,24 +704,3 @@ class FunctionCreationStatus:
                 )
         else:
             self.status_row.finish(f"Created function {self.tag}.")
-
-
-@contextlib.contextmanager
-def enable_output(show_progress: bool = True) -> Generator[None, None, None]:
-    """Context manager that enable output when using the Python SDK.
-
-    This will print to stdout and stderr things such as
-    1. Logs from running functions
-    2. Status of creating objects
-    3. Map progress
-
-    Example:
-    ```python
-    app = modal.App()
-    with modal.enable_output():
-        with app.run():
-            ...
-    ```
-    """
-    with OutputManager.enable_output(show_progress):
-        yield
