+++ modal/requirements/2024.10.txt
@@ -14,13 +14,10 @@ h2==4.1.0
 hpack==4.0.0
 hyperframe==6.0.1
 idna==3.6
-markdown-it-py==3.0.0
-mdurl==0.1.2
 multidict==6.0.5
 protobuf==4.25.3
 pydantic==2.6.4
 pydantic_core==2.16.3
-Pygments==2.17.2
 python-multipart==0.0.9
 sniffio==1.3.1
 starlette==0.36.3
